Implement simpleOrchestrator and rename simpleAction to simpleMutator
